Reorganize endpoints into api.py, add object-counts endpoint
